Add this behind a rule, this directory might not always exist and if it fails then the entire build is broken

+++ Makefile
@@ -429,7 +429,7 @@ obj/%.o: %.c
 clean:
 	-make -C $(LUADIR) clean
 	-rm -f $(OBJS) $(TARGET)
-	-rm -rf obj
+	if [ -d "obj" ]; then rm -rf obj; fi
 
 test: all
 	retroarch --verbose -L lutro_libretro.so test/main.lua
